+++ App.js
@@ -1,25 +1,9 @@
 import * as React from 'react'
 import { Provider as PaperProvider } from 'react-native-paper'
-import { StatusBar } from 'expo-status-bar'
-import { View, StyleSheet } from 'react-native'
 
-import {
-  AuthenticationProvider,
-  useAuthenticationContext,
-} from './context/authentication'
-import Auth from './components/Auth'
-import Home from './components/Home'
+import { AuthenticationProvider } from './context/authentication'
+import Main from './Main'
 import theme from './defaultTheme'
-
-function Main() {
-  const { user } = useAuthenticationContext()
-  return (
-    <View style={styles.container}>
-      {user ? <Home user={user} /> : <Auth />}
-      <StatusBar style="auto" />
-    </View>
-  )
-}
 
 export default function App() {
   return (
@@ -30,11 +14,3 @@ export default function App() {
     </AuthenticationProvider>
   )
 }
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-})
